completed opentypefeature, rename opentype > opentypefeature
