--- spec/visual/samples/opentype
@@ -1,38 +1,0 @@
-table name {
-  nameid 0 "Copyright Enschede";
-  nameid 9 'Jan van Krimpen';
-} name;
-
-
-
-include (../../family.fea)
-
-languagesystem DFLT dflt;
-languagesystem latn dflt;
-
-@className = [A B C.alt];
-@className2 = [a b c.alt];
-
-# some comment
-
-lookup SomeGreatNameHere {
-  sub colon parenright by smile;
-} SomeGreatNameHere;
-
-feature mkmk {
-    include (mkmk.fea);
-} mkmk;
-
-feature liga {
-  sub f i by f_i;
-  # sub f t' by t.liga; # FIRST FIX contextual rule highlight...
-  ignore sub w o r d s exclam;
-} liga;
-
-table OS/2 {
-  Panose 2 4 8 3 5 4 5 2 2 4;
-  FSType 0;
-  # NOTE: For cross-platform consistency, OS/2.TypoAscender and |OS/2.TypoDescender| must add up to the font's UPM value
-  TypoAscender 730;   # Must be the same on all fonts for cross-platform consistency
-  TypoDescender -270; # Must be the same on all fonts for cross-platform consistency
-} OS/2;
